% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tattr.R
\name{add_tattr}
\alias{add_tattr}
\title{add_tattr}
\usage{
add_tattr(
  tob,
  x = NULL,
  xattr = NULL,
  attr = "attr",
  taxon = "taxon",
  tadd = NULL
)
}
\arguments{
\item{tob}{A t* object to add attributes to}

\item{x}{a vector of element names in tob. Alternatively
a dataframe with element names and attribute values as its
columns}

\item{xattr}{a vector comprising the attributes
of the elements in x. This must be the same
length as x}

\item{attr}{If x is a dataframe, a vector specifying the
attribute value columns in x}

\item{taxon}{If x is a dataframe, a vector specifying the
taxon name columns in x}

\item{tadd}{If x is not specified, tadd can be supplied as a t*
object, whose attributes will be added to the corresponding
element names in tob}
}
\value{
An object of the same class as tob, with updated
attributes
}
\description{
Function to add attributes to the vertices of a
t* object
}
