% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tgraph.R
\name{as_tgraph}
\alias{as_tgraph}
\title{as_tgraph}
\usage{
as_tgraph(x)
}
\arguments{
\item{x}{A tvertseq object}
}
\value{
A tgraph object
}
\description{
Function to convert a tvertseq object to a tgraph object
This function simply takes the graph portion of the
tvertseq as the graph for the tgraph, nullifying the
relevance of parent or child relationships
}
