% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tvertseq.R
\name{as_tvertseq}
\alias{as_tvertseq}
\title{as_tvertseq}
\usage{
as_tvertseq(x, mode = c("parent", "child"))
}
\arguments{
\item{x}{A tgraph object}

\item{mode}{A character vector of length one specifying
whether parent or child relationships will be recorded
in the tvertseq. As only highest or lowest rank
conversion is currently supported, mode 'parent' will
result in conversion of the lowest rank, while mode
'child' will result in conversion of the highest rank}
}
\value{
A tvertseq object
}
\description{
Function to coerce a tgraph object to a tvertseq object.
Currently, only the highest ranked or lowest ranked
elements can be coerced to focal elements in a tvertseq,
depending on the argument specification, although
future versions may be adapted to allow any single rank
to be made focal, with parent or child relationships
retained in the tvertseq.
}
