% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_duplicates_del.R
\name{get_duplicates_del}
\alias{get_duplicates_del}
\title{get_duplicates_del}
\usage{
get_duplicates_del(x, ranks = NULL)
}
\arguments{
\item{x}{A tgraph object}

\item{ranks}{An optional numeric or character vector
indicating which ranks in x are to be checked for
duplicates}
}
\value{
A tvertseq with the duplicate classifications
as focal elements
}
\description{
Function to detect elements with multiple higher
classifications in a tgraph object. Detection is
performed simply by calculating the per node degree
of each node in the graph representation of the
elements and their relationships. Degree > 1 indicates
duplicate classifications
}
