% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tattr.R
\name{get_tattr}
\alias{get_tattr}
\title{get_tattr}
\usage{
get_tattr(x)
}
\arguments{
\item{x}{A t* object}
}
\value{
A dataframe of the vertex names in x, along
with their attributes
}
\description{
Function to retrieve the vertex attributes of a t*
object
}
