% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_taxa.R
\name{get_taxa}
\alias{get_taxa}
\title{get_taxa}
\usage{
get_taxa(x, taxa, mode = c("parent", "child"), step = NULL)
}
\arguments{
\item{x}{A tgraph object}

\item{taxa}{The names of the elements whose relationships
will be retrived, or the numeric ids of nodes in the
graph portion of the tgraph object (useful for
retrieving relationships where the name name of a node
may not be known)}

\item{mode}{The directionality of the relationships
to retrieve}

\item{step}{A positive integer specifinyg the
neighbourhood of the relationships to retrive. Specifying
a number greater than the number of ranks in the tgraph
will not cause a failure, and will instead retrieve
all relationships in the direction specified in mode}
}
\value{
a tvertseq object containing the focal elment
and its relationships, as specified by mode
}
\description{
Function to retrive the parent or child relationships
for a set of elements in a tgraph object. The length
of edges beyond which relationships will not be returned
can be set in the function arguments
}
