% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GTS2020.R
\docType{data}
\name{GTS2020}
\alias{GTS2020}
\title{GTS2020}
\format{
An object of class \code{data.frame} with 1536 rows and 10 columns.
}
\usage{
data(GTS2020)
}
\description{
Lookup table of chronostratigraphic interval names
used in the PBDB, the interval type (e.g. ICS
stage, NALMA, ammonoid zone), the geographic extent
of the zone (global in the case of ICS intervals,
regional for other timescales), the base and top
ages (FAD and LAD) of each interval in millions of
years ago. These absolute ages are based primarily
on A Geologic Timescale 2020 by Gradstein et al.
All ICS intervals have GTS age boundaries. Other
timescales have GTS age boundaries either stated
directly in GTS2020, a boundary taken through
correlation with the base or top of an ICS interval
with GTS dates or measured on the correlation charts
in GTS2020. For a minority of intervals, other
literature is used to help map an interval to a
chronostratigraphic unit named in GTS 2020. In the
case of Land Mammal Ages, GTS2020 directs the user
to GTS2012, so those dates are retained
}
\references{
See dataset
}
\keyword{datasets}
