% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_fam.R
\docType{data}
\name{keep_fam}
\alias{keep_fam}
\title{keep_fam}
\format{
An object of class \code{character} of length 71.
}
\usage{
data(keep_fam)
}
\description{
A very small helper dataset recording some specific edits
to the PBDB. Edits are the result of a visual inspection
of potential spelling errors flagged by spell_check. They
are recorded here so that they can be consistently
reapplied to the PBDB, rather than running the function
and using indices to make the edits, as indices are prone
to change with addition of new data to the PBDB
}
\keyword{datasets}
