% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lca.R
\name{lca}
\alias{lca}
\title{lca}
\usage{
lca(graph, ...)
}
\arguments{
\item{graph}{A tgraph object}

\item{...}{A vector of vertex names or numbers for which
the common ancestor will be derived}
}
\value{
The name of the common element
}
\description{
Function to retrieve the last common ancestor of a set
of elements in a tgraph object
}
