% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdb_kingdoms.R
\docType{data}
\name{pbdb_kingdoms}
\alias{pbdb_kingdoms}
\title{pbdb_kingdoms}
\format{
An object of class \code{list} of length 3.
}
\usage{
data(pbdb_kingdoms)
}
\description{
Helper list of the kingdom-level assignment of
phylum names in the PBDB. Used to build a kingdom
column in a PBDB download if taxonomy fields were
requested. List of three character vectors, named
animals, plants, protists. There is only a single
fungal phylum in the database so this is assigned
directly, while all other phyla are assigned to
bacteria
}
\keyword{datasets}
