% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_t.R
\name{plot_t}
\alias{plot_t}
\title{plot_t}
\usage{
plot_t(x, taxa = NULL, taxa2 = NULL, mode = c("parent", "child"), step = NULL)
}
\arguments{
\item{x}{a tgraph or tvertex object}

\item{taxa}{A character vector of element names whose
relationships will be plotted}

\item{taxa2}{An optional numeric vector of element names
whose relationships will be plotted}

\item{mode}{The direction of the relationships to be
plotted}

\item{step}{A positive integer specifinyg the
neighbourhood of the relationships to plot. Specifying
a number greater than the number of ranks in the t* object
will not cause a failure, and will instead plot all
relationships in the direction specified in mode}
}
\value{
A plot of the relationships of the specified
elements
}
\description{
Function to plot the parent or child relationships of an
element in a tgraph or tvertseq object. The length
of edges beyond which relationships will not be plotted
can be set in the function arguments
}
