% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_t2.R
\name{plot_t2}
\alias{plot_t2}
\title{plot_t2}
\usage{
plot_t2(
  x,
  taxon,
  trank,
  ranks,
  mode = c("parent", "child", "all"),
  step = NULL
)
}
\arguments{
\item{x}{a dataframe containing hierarchically organised
data in columns}

\item{taxon}{A character vector of element names whose
relationships will be plotted (these must be of the same
rank)}

\item{trank}{A character vector of length one corresponding
to the column name in x in which taxa is located}

\item{ranks}{A character vector corresponding to the column
names in x, given in hierarchical order}

\item{mode}{The direction of the relationships to be
plotted}

\item{step}{A positive integer specifinyg the
neighbourhood of the relationships to plot. Specifying
a number greater than the number of ranks in the t* object
will not cause a failure, and will instead plot all
relationships in the direction specified in mode}
}
\value{
A plot of the relationships of the specified
elements
}
\description{
Function to plot the parent or child relationships of an
element in a hierarchically organised dataframe. An
accompaniment to @seealso plot_t which internally coerces
the input into a format that would be plotted as in
plot_t. The length of edges beyond which relationships
will not be plotted can be set in the function arguments
}
