% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tgraph_tedgelist.R
\name{tgraph_tedgelist}
\alias{tgraph_tedgelist}
\title{tgraph_tedgelist}
\usage{
tgraph_tedgelist(from)
}
\arguments{
\item{from}{a tgraph object}
}
\value{
a tgraph object with the alternative
formatting of the element relationships in from
}
\description{
Function to convert the relationships in a tgraph
from a graph format to an edgelist. The current
format of the supplied object is detected and
automatically converted to the other format
}
