% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tvertseq.R
\name{tvertseq}
\alias{tvertseq}
\title{tvertseq}
\usage{
tvertseq(
  x,
  hrank,
  taxon = "taxon",
  link = "link",
  ranks = "rank",
  lranks = "lrank",
  mode = c("parent", "child")
)
}
\arguments{
\item{x}{A four column dataframe consisting of two pairs of
columns. The first column specifies the focal elements in
the tvertseq object. The second column specifies elements
related, row by row, to the elements in column one. This
first pair of columns is effectively an edgelist. The third
and fourth columns are the respective ranks of the
corresponding elements in columns one and two}

\item{hrank}{A named vector of consective integers starting
from one. One is the highest rank in any t* object. The
names in hrank are the names of the actual rank system.}

\item{taxon}{The name of the focal element column in x}

\item{link}{The name of related elements column in x}

\item{ranks}{The name of the column denoting the ranks of
the focal elements in x}

\item{lranks}{The name of the column denoting the ranks of
the related elements in x}

\item{mode}{The direction of the relationships in the focal
and related elements column. If parent, then the
relationship elements exist at a higher rank than their
corresponding focal elements, and vice versa if child}
}
\value{
A tvertseq object
}
\description{
Function to create a tvertseq object from scratch. The
tvertseq object displays the higher or lower relationships
for a given set of elements. These relationships may
intersect between elements and elements need not be at the
same level. The function can be used directly, but
requires more inputs compared to tvertex or tgraph and
exists primarily to complete the logical hierarchy between
these objects, and as a widely used carrier internally
within t* functions
}
